added a delete user button to the user admin panel

+++ apps/web/src/actions/adminActions/adminActions.js
@@ -5,7 +5,13 @@ async function adminAction({ request }) {
     const fd = await request.formData();
     const data = Object.fromEntries(fd);
     try {
-        await api({ url: `/posts/${data.postId}`, options: { method: 'DELETE' } })
+        if (data.intent === 'deletePost') {
+            await api({ url: `/posts/${data.postId}`, options: { method: 'DELETE' } });
+        }
+
+        if (data.intent === 'deleteUser') {
+            await api({ url: `/users/${data.userId}`, options: { method: 'DELETE' } });
+        }
     } catch (error) {
 
     }
